@@ -25,7 +25,7 @@ plugins {
 // DO NOT FORGET TO DOCUMENT CHANGES IN CHANGELOG.md
 //
 // Add a GitHub release for every new release: https://github.com/otto-de/edison-microservice/releases
-def edison_version = "2.7.7"
+def edison_version = "3.0.0-SNAPSHOT"
 //
 //
 //
@@ -78,8 +78,8 @@ subprojects {
         }
     }
 
-    targetCompatibility = '11'
-    sourceCompatibility = '11'
+    targetCompatibility = '17'
+    sourceCompatibility = '17'
 
     project.afterEvaluate {
         tasks.withType(JavaCompile) {
