@@ -1,7 +1,7 @@
 package de.otto.edison.validation.validators;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 import java.time.Instant;
 import java.time.format.DateTimeParseException;
 
