@@ -1,9 +1,8 @@
 package de.otto.edison.health.indicator;
 
+import jakarta.validation.constraints.Min;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
-
-import javax.validation.constraints.Min;
 
 /**
  * Properties used to configure the graceful shutdown of Edison microservices.
