@@ -1,7 +1,6 @@
 package de.otto.edison.togglz;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
@@ -16,14 +15,8 @@ import static org.springframework.boot.SpringApplication.run;
 @SpringBootApplication
 public class TestServer {
 
-    private static ApplicationContext ctx;
-
-    public static ApplicationContext applicationContext() {
-        return ctx;
-    }
-
     public static void main(String[] args) {
-        ctx = run(TestServer.class, args);
+        run(TestServer.class, args);
     }
 
 }
